chore: Following subcommand pattern for magento cli
**Tickets:**
* [sc-82477](https://app.shortcut.com/publicsq/story/82477)
