+++ PublicSquare/Payments/etc/di.xml
@@ -149,7 +149,7 @@
     <type name="Magento\Framework\Console\CommandListInterface">
         <arguments>
             <argument name="commands" xsi:type="array">
-                <item name="psq" xsi:type="object">PublicSquare\Payments\Cli\PsqCli</item>
+                <item name="psq" xsi:type="object">PublicSquare\Payments\Cli\PsqConfigureWebhooksCommand</item>
 
             </argument>
         </arguments>
